+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Replace header toggle with option bar icon (Thanks to Fr33Lancer for the icon)
 - Add button to select all launchers (for area attack)
 - Set unready state for area attack
+- Handle Galactic War spec_tag
 
 ## 1.0.0